Primed for NPM and require.
